+++ src/biodwh2-drugcentral/src/main/java/de.unibi.agbi.biodwh2.drugcentral/model/AtcDdd.java
@@ -2,8 +2,8 @@ package de.unibi.agbi.biodwh2.drugcentral.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphNodeLabel;
+import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 
 @SuppressWarnings("unused")
 @JsonPropertyOrder({"id", "atc_code", "ddd", "unit_type", "route", "comment", "struct_id"})
@@ -27,5 +27,5 @@ public final class AtcDdd {
     @GraphProperty("comment")
     public String comment;
     @JsonProperty("struct_id")
-    public Integer structId;
+    public Long structId;
 }
