+++ src/biodwh2-drugcentral/src/main/java/de.unibi.agbi.biodwh2.drugcentral/model/Pdb.java
@@ -2,8 +2,8 @@ package de.unibi.agbi.biodwh2.drugcentral.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphNodeLabel;
+import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.drugcentral.etl.DrugCentralGraphExporter;
 
 @SuppressWarnings("unused")
@@ -16,7 +16,7 @@ public final class Pdb {
     @JsonProperty("id")
     public String id;
     @JsonProperty("struct_id")
-    public Integer structId;
+    public Long structId;
     @JsonProperty("pdb")
     @GraphProperty("pdb")
     public String pdb;
