+++ src/biodwh2-drugcentral/src/main/java/de.unibi.agbi.biodwh2.drugcentral/model/Pka.java
@@ -9,7 +9,7 @@ public final class Pka {
     @JsonProperty("id")
     public Integer id;
     @JsonProperty("struct_id")
-    public Integer structId;
+    public Long structId;
     @JsonProperty("pka_level")
     public String pkaLevel;
     @JsonProperty("value")
