+++ src/biodwh2-drugcentral/src/main/java/de.unibi.agbi.biodwh2.drugcentral/model/Struct2Parent.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 @JsonPropertyOrder({"struct_id", "parent_id"})
 public final class Struct2Parent {
     @JsonProperty("struct_id")
-    public Integer structId;
+    public Long structId;
     @JsonProperty("parent_id")
-    public Integer parentId;
+    public Long parentId;
 }
