+++ src/biodwh2-drugcentral/src/main/java/de.unibi.agbi.biodwh2.drugcentral/model/Structure.java
@@ -2,8 +2,8 @@ package de.unibi.agbi.biodwh2.drugcentral.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphNodeLabel;
+import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 
 @SuppressWarnings("unused")
 @JsonPropertyOrder({
@@ -24,7 +24,7 @@ public final class Structure {
     public String cdMolweight;
     @JsonProperty("id")
     @GraphProperty("id")
-    public int id;
+    public Long id;
     @JsonProperty("clogp")
     @GraphProperty("clogp")
     public String clogp;
